requirements.txt
click==8.0.3
colorama==0.4.4
et-xmlfile>=1.1.0
Flask>=2.0.2
Flask-Login>=0.5.0
Flask-SQLAlchemy>=2.5.1
Flask-WTF==1.0.0
geopy>=2.2.0
gunicorn==20.1.0
numpy>=1.21.4
openpyxl>=3.0.9
pandas>=1.3.4
python-dateutil>=2.8.2
protobuf==3.19.1
requests>=2.26.0
idna>=3.3
retry>=0.9.2
six>=1.16.0
SQLAlchemy>=1.4.28
tomli>=1.2.2
typing_extensions>=4.0.1
urllib3>=1.26.7
Werkzeug>=2.0.2